last-with-explanation

+++ src/controllers/libroController.js
@@ -1,37 +1,35 @@
 const Libro = require("../models/libroModel");
 
-exports.getAllLibros = async (req, res) => {
+exports.getAllLibros = async (req, res, next) => {
   try {
     const libros = await Libro.find();
     res.status(200).json(libros);
   } catch (error) {
-    res.status(500).json({ error: "Error al obtener los libros" });
+    next(error);
   }
 };
 
-exports.getLibroById = async (req, res) => {
+exports.getLibroById = async (req, res, next) => {
   try {
     const libro = await Libro.findById(req.params.id);
-    //if (!libro) {
-     // return res.status(404).json({ error: "Libro no encontrado" });
-    //}
+    
     res.status(200).json(libro);
   } catch (error) {
-    res.status(500).json({ error: "Error al obtener el Libro" });
+    next(error);   
   }
 };
 
-exports.createLibro = async (req, res) => {
+exports.createLibro = async (req, res, next) => {
   try {
     const nuevoLibro = await Libro.create(req.body);
     await nuevoLibro.save();
     res.status(201).json(nuevoLibro);
   } catch (error) {
-    res.status(500).json({ error: "Error al crear el Libro" });
+    next(error);
   }
 };
 
-exports.updateLibro = async (req, res) => {
+exports.updateLibro = async (req, res, next) => {
   try {
     const libro = await Libro.findByIdAndUpdate(req.params.id, req.body, {
       new: true,
@@ -43,18 +41,22 @@ exports.updateLibro = async (req, res) => {
 
     res.status(200).json(libro);
   } catch (error) {
-    res.status(500).json({ error: "Error al actualizar el Libro" });
+    next(error);
   }
 };
 
-exports.deleteLibro = async (req, res) => {
+exports.deleteLibro = async (req, res, next) => {
   try {
     const libroId = req.params.id;
-
+    if(!libroId){
+      const error = new error("Libro no encontrado");
+      error.status = 401;
+      throw(error);
+    }
     const libroEliminado = await Libro.findByIdAndRemove(libroId);
 
     res.status(200).json(libroEliminado);
   } catch (error) {
-    res.status(500).json({ error: "Error al eliminar el Libro" });
+    next(error);
   }
 };
